refactor(roadmap): split plan_and_claim and thin the command surface
Break plan_and_claim into blocked/resolve/claim phases, move evented
writes into mutations, relocate merge_hold_gate onto merge_sweep, and
extract Tauri commands so mod.rs is only module wiring.
